Move even more stuff into seperate files
