+++ WindowsTrojan/Source/Utils/Random.c
@@ -1,0 +1,13 @@
+#include "../MEMZ.h"
+
+HCRYPTPROV prov;
+
+int random() {
+	if (prov == NULL)
+		if (!CryptAcquireContext(&prov, NULL, NULL, PROV_RSA_FULL, CRYPT_SILENT | CRYPT_VERIFYCONTEXT))
+			ExitProcess(1);
+
+	int out;
+	CryptGenRandom(prov, sizeof(out), (BYTE *)(&out));
+	return out & 0x7fffffff;
+}
