+++ WindowsTrojan/Source/Utils/Utils.h
@@ -1,0 +1,3 @@
+int random();
+void strReverseW(LPWSTR str);
+LRESULT CALLBACK msgBoxHook(int, WPARAM, LPARAM);
